requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twilio==8.10.0
supabase==2.1.0
google-generativeai==0.3.1
google-api-python-client>=2.115.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
dateparser==1.2.0
httpx==0.24.1 